+++ src/Button.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Button({ children }) {
+  return (
+    <button type="button">
+      {children}
+    </button>
+  );
+}
+
+export default Button;
