+++ src/index.jsx
@@ -1,76 +1,6 @@
 import React, { useState } from 'react';
 import ReactDOM from 'react-dom';
-
-function Counter({ count, onClick }) {
-  return (
-    <button type="button" onClick={onClick}>
-      Click me!
-      (
-      {count}
-      )
-    </button>
-  );
-}
-
-function Button({ children }) {
-  return (
-    <button type="button">
-      {children}
-    </button>
-  );
-}
-
-function Buttons({ onClick }) {
-  return (
-    <p>
-      {[1, 2, 3, 4, 5].map((i) => (
-        <Button key={i} onClick={onClick}>
-          {i}
-        </Button>
-      ))}
-    </p>
-  );
-}
-
-function Page({ count, onClick }) {
-  return (
-    <div>
-      <p>Hello,world!</p>
-      <p>Hi!</p>
-      <Counter
-        count={count}
-        onClick={onClick}
-      />
-      <Buttons
-        count={count}
-        onClick={onClick}
-      />
-    </div>
-  );
-}
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-    countOne: 1,
-  });
-
-  const { count } = state;
-
-  function handleClick() {
-    setState({
-      count: count + 1,
-      countOne: count + 1,
-    });
-  }
-
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-    />
-  );
-}
+import App from './App';
 
 ReactDOM.render(
   <App />,
